+++ src/OpenLoco/src/Vehicles/Orders.h
@@ -161,5 +161,5 @@ namespace OpenLoco::Vehicles
 
 #pragma pack(pop)
 
-    uint16_t reverseVehicleOrderTable(uint32_t tableOffset, uint32_t orderOfInterest);
+    uint16_t reverseVehicleOrderTable(uint16_t tableOffset, uint16_t orderOfInterest);
 }
